fix: resolve tsc --build circular dependency (db ↔ middleware)
Move RateLimitMiddlewareConfig interface from @ai-learning-hub/middleware
to @ai-learning-hub/types to break the circular project reference:
db imported middleware for the type, but middleware references db.

Both packages already reference types, so this eliminates the cycle.
Middleware re-exports the type for backward compatibility.
